[Move] notebook from `src/code/common` to `src/code/platform`
